Fix import paths to revert internal rename.
